+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [Unreleased]
+
 ## [0.0.3] - 2026-07-02
 
 ### Fixed
@@ -26,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Comprehensive documentation
 - Unit test suite
 
+[Unreleased]: https://github.com/jhotadhari/release-kit/compare/v0.0.3...HEAD
 [0.0.3]: https://github.com/jhotadhari/release-kit/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/jhotadhari/release-kit/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/jhotadhari/release-kit/releases/tag/v0.0.1